+++ src/tokens/components/hero.ts
@@ -1,0 +1,4 @@
+import hero from '../../../tokens/component/hero.json';
+import { flattenTokens } from '../utils';
+
+export const heroTokens = flattenTokens(hero as never);